Corrected problems with board
Changes:
1. Made NUMCOLS and NUMROWS properties of board. Makes it easier to add
the ability to dynamically add rows and columns later.
2. Clicking an occupied tile on board no longer puts an "undefined"
tile in the hand. Need to be careful with giving functions correct
inputs. (GAME.addTileToHand was given a string, it takes an object
literal. Hence a bug.)
3. Made it so that board tiles only highlight and change when you have
selected a tile in your hand to put on the board.

Known bugs:
1. It takes forever for the initial tiles to be placed in the player's
hand. Not sure where the code is slowing down or where call could be
delayed.
2. Alert says 5 seconds until game starts.  Creating an alert actually
suspends Javascript execution and game actually doesn't start until ok
is pressed. Left it in for now but we need to look into using a more
suitable UI element for this.
3. Not really a bug but board is fixed size and restricts the way users
can arrange words. Makes gameplay difficult/impossible. Looking to make
changes on this front in my next version.